Update linq.js
Added FilteredEnumerable, to make it possible to chain boolean conditions. An example use case could be filtering columns to show in a table.
myReport = myReport.Where(); // this is a way to set up the boolean chaining. Normally you would pass in a boolean condition right into .Where
for(var x = 0; x < players.length; x++){
   var player = players[x];
   myReport = myReport.Or(row => row.player === player);
}

The performance needs to be tested, but this is an easy alternative way of writing: myReport = myReport.Where(x=>players.indexOf(x.player) > -1);