Update linq.js
Added in support for LeftJoin, RightJoin, and FullJoin. In order to use each one, you must pass in an additional predicate specifying how to handle the unmatched item.

I.e for LeftJoin, the final argument will be something like (leftMiss) => ({TypeOfMiss: "Left", Object: leftMiss}) . And that is what will be returned for each object from the left that didn't have a match on the right.