Update linq.js
Added more robust functionality to Joins. Joins behave much more like SQL joins now, where the join condition is not a KEY but a boolean expression, and the return result of a join is a Joining object with a Left and Right property, very much like a real SQL join behaves. If performing a Left/Right/Full join (as opposed to an (Inner)Join), then any rows not passing the join condition will be set to null in the Joining object. Optionally, a Select predicate may be supplied as a 3rd parameter which specifies how to treat the join, in terms of what data should be returned for that row.