Update linq.js
Improved Sequence methods, added Scan method.

Scan method applies an aggregating accumulation on the Enumerable, so if you had internal array [1,2,3] and applied an accumulator like a,b => a+b and called ToArray after, you'd end up with [1,3,6]. There is an optional argument to provide a seed value.

The updated Sequence methods allow you to create a brand new Enumerable, or inject a new sequence into the current Enumerable lazily. The callback method passed in as the 'generator' param has a signature of currentArray, iterationIndex. Example usage to generate the Fibonacci sequence

_.Sequence(10,[],function(curr,i){
   if(i <= 1){
       return i;
   }
   return curr[i-1] + curr[i-2]
}).ToArray()

=====> [0, 1, 1, 2, 3, 5, 8, 13, 21, 34]