Optimize Push/Pop
Refactored the 'foreachactionstack' code. Able to intercept each ForEachActionPredicate and modify its value. PushPredicate inherits from ForEachActionPredicate, and every call to Push checks if the latest ForEachActionPredicate is a PushPredicate. If so, it just adds to that predicates internal array. Otherwise it pushes a new PushPredicate to the stack. Basically same principle with PopPredicate